+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - KNN graph function.
+- Allow cluster to accept adjacency matrix as represented by numpy arrays.
 
 ## [v0.1.3] 2024-10-28